Move node related code to `node` module
