+++ node/src/main/scala/memeid/node/Node.scala
@@ -1,12 +1,9 @@
 package memeid.node
 
-import java.net.{InetAddress, NetworkInterface}
+import memeid.digest.Digest
+import memeid.bits.fromBytes
 
-import scala.collection.JavaConverters._
 import scala.util.Random
-
-import memeid.bits._
-import memeid.digest._
 
 trait Node {
 
@@ -92,28 +89,3 @@ object Node {
 
 }
 
-@SuppressWarnings(Array("scalafix:Disable.toString"))
-object Sys {
-
-  def getNetworkInterfaces: Set[String] = {
-    val localHost     = InetAddress.getLocalHost
-    val hostName      = localHost.getCanonicalHostName
-    val baseAddresses = Set(localHost.toString, hostName)
-    NetworkInterface.getNetworkInterfaces.asScala.foldLeft(baseAddresses)({
-      case (addrs, ni) =>
-        addrs ++ ni.getInetAddresses.asScala.map(_.toString).toSet
-    })
-  }
-
-  def getLocalInterfaces: Set[String] = {
-    val localHost = InetAddress.getLocalHost
-    val hostName  = localHost.getCanonicalHostName
-    InetAddress.getAllByName(hostName).map(_.toString).toSet
-  }
-
-  def getProperties: Map[String, String] = {
-    val props = System.getProperties
-    val keys  = props.stringPropertyNames
-    keys.asScala.map(k => k -> props.getProperty(k)).toMap
-  }
-}
